+++ packages/examples/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ['../../.eslintrc.js', 'plugin:react/recommended'],
+  rules: {
+    'react/prop-types': 0,
+    'jsx-a11y/anchor-is-valid': 0,
+  },
+};