+++ packages/layout/src/node/setAlign.js
@@ -1,4 +1,4 @@
-import Yoga from '@react-pdf/yoga';
+import Yoga from 'yoga-layout/sync';
 import { upperFirst } from '@react-pdf/fns';
 
 const ALIGN = {