+++ packages/yoga/index.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-unresolved
-import initYoga from 'yoga-wasm-web/asm';
-
-export default initYoga();